requirements.txt
streamlit
pandas
numpy
psycopg2-binary
google-genai
pydantic
tenacity